ipl-2025-pressconference-rag-bot/requirements.txt
anthropic==0.40.0
chromadb==0.5.23
sentence-transformers==3.3.1
streamlit==1.40.2
beautifulsoup4==4.12.3
requests==2.32.3
tiktoken==0.8.0
python-dotenv==1.0.1
